backend_python/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
google-cloud-firestore==2.17.2
google-cloud-storage==2.18.2
firebase-admin==6.5.0
reportlab==4.2.2
qrcode[pil]==7.4.2
pydantic-settings==2.5.2
pytest==8.3.3
